app/requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-mcp-adapters
langchain-community
langgraph
fastmcp
dashscope
gradio